refactor(hooks): extract phase runner into ui base
