+++ crates/dbflux_ui_base/Cargo.toml
@@ -24,6 +24,7 @@ uuid.workspace = true
 log.workspace = true
 anyhow = { workspace = true }
 chrono.workspace = true
+serde_json.workspace = true
 
 [dev-dependencies]
 tempfile = "3"
