+++ crates/dbflux_ui_base/src/lib.rs
@@ -10,6 +10,7 @@ pub mod app_state_entity;
 pub mod async_ext;
 pub mod dashboard_manager;
 pub mod file_dialog;
+pub mod hook_phase_runner;
 pub mod keymap;
 pub mod modal_frame;
 pub mod modals;
